+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## Current Release 
+### 0.211.0 
+**Release Date:** Wed Sep  5 15:19:28 UTC 2018     
+## Previous Releases 
 ### 0.210.0 
 **Release Date:** Tue Sep  4 14:28:35 UTC 2018     
-## Previous Releases 
 * Feature - data cleaning history api
 ### 0.209.0 
 **Release Date:** Fri Aug 31 02:33:32 UTC 2018